add quaternion_operation .so
